pyproject.toml
[project]
name = "slack-ai-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.0",
    "bs4>=0.0.2",
    "fastapi>=0.118.2",
    "google-cloud-aiplatform>=1.120.0",
    "google-cloud-logging>=3.12.1",
    "google-cloud-secret-manager>=2.24.0",
    "google-cloud-storage>=2.18.0",
    "langchain>=1.0.0",
    "langchain-community>=0.4",
    "langchain-google-community>=3.0.0",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.0",
    "langgraph-checkpoint-postgres>=3.0.0",
    "psycopg[binary]>=3.2.12",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "slack-bolt>=1.26.0",
    "uvicorn>=0.37.0",
    "watchfiles>=1.1.1",
    "yoyo-migrations>=9.0.0",
    "ruff>=0.14.9",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.ruff]
# 基本設定
line-length = 88  # 1行の最大文字数
target-version = "py313"  # Python 3.13を対象
fix = true  # 自動修正を有効化

# ファイルの除外設定
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "*.pyc",
    ".pytest_cache",
    ".mypy_cache",
    "migrations",
    "build",
    "dist",
]

[tool.ruff.lint]
# 有効にするルール
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort（import順序）
    "N",    # pep8-naming
    "UP",   # pyupgrade（古い構文を最新に）
    "B",    # flake8-bugbear（バグになりやすいコード）
    "SIM",  # flake8-simplify（簡潔なコード）
    "C4",   # flake8-comprehensions（内包表記の最適化）
    "DTZ",  # flake8-datetimez（datetime）
    "T20",  # flake8-print（print文の検出）
    "RET",  # flake8-return（return文の改善）
    "PTH",  # flake8-use-pathlib（pathlibの使用推奨）
    "ERA",  # eradicate（コメントアウトされたコード）
    "PL",   # Pylint
    "RUF",  # Ruff独自ルール
]

# 無視するルール
ignore = [
    "E501",   # line-too-long（line-lengthで制御）
    "B008",   # デフォルト引数での関数呼び出し（FastAPIで使用）
    "N815",   # 変数名の大文字小文字混在（定数で使用）
    "PLR0913", # 引数が多すぎる（5個まで許可）
]

# ファイル別の設定
[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]  # __init__でのimport
"tests/*" = ["T20", "PLR2004"]    # テストでのprint文と定数比較を許可

[tool.ruff.lint.isort]
# import順序の設定
known-first-party = ["src"]
section-order = [
    "future",
    "standard-library",
    "third-party",
    "first-party",
    "local-folder"
]

[tool.ruff.lint.pydocstyle]
# docstringのスタイル（Google style）
convention = "google"

[tool.ruff.format]
# フォーマット設定
quote-style = "double"        # ダブルクォート使用
indent-style = "space"        # スペースでインデント
skip-magic-trailing-comma = false  # 末尾カンマを保持
line-ending = "auto"          # 改行コードは自動検出

